Rule the editor at 80 and 118, bar what changed
- Page guides under the text at both columns, measured from the
  font's own advance and skipped entirely for a face that is not
  fixed pitch, so a rule never lands between characters.
- Uncommitted lines take a change bar down the gutter's inner edge,
  where it sits against the code it belongs to. It replaces the
  tinted line number, which said the same thing but could not be
  read down a scrolling file.
- The ruler and its line arithmetic move to their own file, which
  is what the editor's length limit had left no room for.
