+++ README.md
@@ -58,6 +58,9 @@ In workflow order; the loop from prompt to review repeats before shipping.
   fact (checks a dot, approval a tick, changes requested a diff, a conflict
   a warning), its checks and queued merges watched until they settle (so
   the sidebar is the dashboard and no two badges look alike)
+- Edits with syntax highlighting, `.editorconfig` indentation, page guides
+  at columns 80 and 118 and a change bar beside every uncommitted line
+  (so the editor says what is yours before the diff does)
 - Marks a worktree whose pane has been holding several cores for ten
   minutes, naming what is running (so one runaway command cannot quietly
   starve every other agent on the machine)
